all test cases; start.go; launcher.sh

+++ src/laucher.sh
@@ -1,17 +1,14 @@
 #!/bin/bash
-#if [ "$#" -ne 2 ]; then
-#    echo Usage : $0 ' <min server id> <max server id>'
-#    exit
-#fi
+if [ "$#" -ne 1 ]; then
+    echo Usage : $0 ' <Config File Number (1~8)>'
+    exit
+fi
 
 pkill server
 rm ../logs/*
-#xterm go run master/master.go
-gnome-terminal -e "go run master/master.go config07.json"
-#for i in `seq $1 $2`; do
+gnome-terminal -e "go run master/master.go config0$1.json"
 for i in `seq 4001 4004`; do
-#    server $i &
-    go run server/server.go $i config07.json &
+    go run server/server.go $i config0$1.json &
 done
-gnome-terminal -e "go run client/client.go 4999 config07.json"
+gnome-terminal -e "go run client/client.go 4999 config0$1.json"
 echo servers launched!
